+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Thu, 01 Aug 2024 16:52:17 GMT and should not be manually modified.
+This log was last generated on Fri, 02 Aug 2024 22:55:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 27.0.1
+
+Fri, 02 Aug 2024 22:55:10 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.1.0
+
+## 27.0.0
+
+Fri, 02 Aug 2024 22:15:53 GMT
+
+### Major changes
+
+- Bump @ni/nimble-components to v31.0.0
 
 ## 26.1.2